# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  role: string;
  campuses?: string[];
  subjects?: string[];
  isActive?: boolean;
  /** @nullable */
  lastLoginAt?: string | null;
}

export interface StudentSearchResult {
  studentId: string;
  studentName: string;
  /** @nullable */
  instituteName?: string | null;
  /** @nullable */
  sectionName?: string | null;
  /** @nullable */
  attendancePct?: number | null;
  presentCount?: number;
  totalCount?: number;
}

export interface StudentOverview {
  studentId: string;
  studentName: string;
  /** @nullable */
  instituteName?: string | null;
  /** @nullable */
  sectionName?: string | null;
  totalSessions: number;
  presentCount: number;
  absentCount: number;
  attendancePct: number;
  inRecovery: boolean;
  coursesInRecovery: number;
}

export interface SubjectAttendance {
  subjectTitle: string;
  present: number;
  total: number;
  pct: number;
  meetsRequirement: boolean;
}

export interface SessionRecord {
  date: string;
  sessionTitle: string;
  subjectTitle: string;
  attendanceStatus: string;
  /** @nullable */
  markingMethod?: string | null;
}

export interface QuizItem {
  subjectTitle: string;
  title: string;
  score: number;
  maxScore: number;
  percentage: number;
  status: string;
  /** @nullable */
  date?: string | null;
}

export interface QuizSummary {
  attempted: number;
  total: number;
  avgPct: number;
}

export interface StudentQuizzes {
  classroomQuizzes: QuizItem[];
  moduleQuizzes: QuizItem[];
  classroomSummary: QuizSummary;
  moduleSummary: QuizSummary;
}

export interface SubjectSummary {
  subjectTitle: string;
  studentCount: number;
  presentCount: number;
  totalCount: number;
  pct: number;
}

export interface CampusSummaryItem {
  instituteName: string;
  studentCount: number;
  sectionCount: number;
  subjectCount: number;
  presentCount: number;
  totalCount: number;
  pct: number;
}

export interface SectionSummaryItem {
  instituteName: string;
  sectionName: string;
  studentCount: number;
  presentCount: number;
  totalCount: number;
  pct: number;
}

export interface DashboardSummary {
  totalStudents: number;
  totalCampuses: number;
  avgAttendancePct: number;
  subjectsBelow80: number;
  subjectBreakdown: SubjectSummary[];
  campusBreakdown: CampusSummaryItem[];
  sectionBreakdown: SectionSummaryItem[];
  needsAttention: StudentSearchResult[];
  updatedAt?: string;
}

export interface DashboardStudent {
  studentId: string;
  studentName: string;
  instituteName: string;
  /** @nullable */
  sectionName: string | null;
  presentCount: number;
  totalCount: number;
  attendancePct: number;
  /** @nullable */
  classroomAvg?: number | null;
  /** @nullable */
  moduleAvg?: number | null;
  spiPath: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  campuses: string[];
  subjects: string[];
  isActive: boolean;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
}

export interface UserInput {
  name: string;
  email: string;
  role: string;
  password: string;
  campuses?: string[];
  subjects?: string[];
  isActive?: boolean;
}

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  password?: string;
  campuses?: string[];
  subjects?: string[];
  isActive?: boolean;
}

export interface Campus {
  id: string;
  name: string;
  /** @nullable */
  instituteId?: string | null;
  createdAt: string;
}

export interface CampusInput {
  name: string;
  instituteId?: string;
}

export interface CampusUpdate {
  name?: string;
  instituteId?: string;
}

export interface RoleMeta {
  value: string;
  label: string;
  description: string;
}

export interface AdminMeta {
  roles: RoleMeta[];
  campuses: string[];
  subjects: string[];
}

export interface BqDataset {
  datasetId: string;
}

export interface BqTable {
  tableId: string;
  kind: string;
}

export type BqPreviewRowsItem = { [key: string]: unknown };

export interface BqPreview {
  columns: string[];
  rows: BqPreviewRowsItem[];
}

export interface ProfileUpdate {
  name?: string;
  currentPassword?: string;
  newPassword?: string;
}

export type SearchStudentsParams = {
  q: string;
  limit?: number;
};

export type FetchStudentOverviewParams = {
  /**
   * SPI share token for public access (used when no staff session)
   */
  t?: string;
};

export type FetchStudentSubjectsParams = {
  /**
   * SPI share token for public access (used when no staff session)
   */
  t?: string;
};

export type FetchStudentRecentSessionsParams = {
  /**
   * SPI share token for public access (used when no staff session)
   */
  t?: string;
};

export type FetchStudentQuizzesParams = {
  /**
   * SPI share token for public access (used when no staff session)
   */
  t?: string;
};

export type GetDashboardStudentsParams = {
  search?: string;
  limit?: number;
  campus?: string;
  section?: string;
  attendanceBand?: string;
};

export type GetDashboardFiltersParams = {
  /** When set, sections are limited to this campus */
  campus?: string;
};

export interface DashboardFilters {
  campuses: string[];
  sections: string[];
  updatedAt: string;
};

export type ListTablesParams = {
  dataset: string;
};

export type PreviewTableParams = {
  dataset: string;
  table: string;
  limit?: number;
};