# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  campuses: string[];
  subjects: string[];
  isActive: boolean;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
}